Rename spawners to factories
