+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Actor/CarlaActorFactoryBlueprint.h
@@ -6,25 +6,25 @@
 
 #pragma once
 
-#include "Carla/Actor/ActorSpawner.h"
 #include "Carla/Actor/ActorSpawnResult.h"
+#include "Carla/Actor/CarlaActorFactory.h"
 
 #include "GameFramework/Actor.h"
 
-#include "ActorSpawnerBlueprint.generated.h"
+#include "CarlaActorFactoryBlueprint.generated.h"
 
-/// Base class for Blueprints implementing AActorSpawner interface.
+/// Base class for Blueprints implementing ACarlaActorFactory interface.
 ///
 /// Blueprints deriving from this class are expected to override
-/// GenerateDefinitions and SpawnActor functions.
+/// GetDefinitions and SpawnActor functions.
 UCLASS(Abstract, BlueprintType, Blueprintable)
-class CARLA_API AActorSpawnerBlueprint : public AActorSpawner
+class CARLA_API ACarlaActorFactoryBlueprint : public ACarlaActorFactory
 {
   GENERATED_BODY()
 
 public:
 
-  TArray<FActorDefinition> MakeDefinitions() final
+  TArray<FActorDefinition> GetDefinitions() final
   {
     return GenerateDefinitions();
   }
