+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Game/CarlaGameModeBase.cpp
@@ -12,8 +12,8 @@
 #include "Game/CarlaPlayerState.h"
 #include "Game/Tagger.h"
 #include "Game/TaggerDelegate.h"
+#include "Sensor/OldSensorFactory.h"
 #include "Sensor/Sensor.h"
-#include "Sensor/SensorFactory.h"
 #include "Settings/CarlaSettings.h"
 #include "Settings/CarlaSettingsDelegate.h"
 #include "Util/RandomEngine.h"
